+++ src/note.jl
@@ -5,15 +5,15 @@ export Note
 Data structure describing a "music note".
 ## Fields:
 * `value::UInt8` : Pitch, starting from C0 = 0, adding one per semitone (middle-C is 60).
-* `duration::UInt64` : Duration in ticks.
-* `position::UInt64` : Position in absolute time (since beginning of track), in ticks.
+* `duration::UInt` : Duration in ticks.
+* `position::UInt` : Position in absolute time (since beginning of track), in ticks.
 * `channel::UInt8` : Channel of the track that the note is played on.
 * `velocity::UInt8` : Dynamic intensity. Cannot be higher than 127 (0x7F).
 """
 type Note
     value::UInt8
-    duration::UInt64
-    position::UInt64
+    duration::UInt
+    position::UInt
     channel::UInt8
     velocity::UInt8
