+++ src/sysexevent.jl
@@ -4,7 +4,7 @@ export SysexEvent
 See `TrackEvent`.
 """
 type SysexEvent <: TrackEvent
-    dT::Int64
+    dT::Int
     data::Array{UInt8,1}
 end
 
@@ -12,7 +12,7 @@ function issysexevent(b::UInt8)
     b == 0xF0
 end
 
-function readsysexevent(dT::Int64, f::IO)
+function readsysexevent(dT::Int, f::IO)
     data = UInt8[]
     read(f, UInt8) # Eat the SYSEX that's on top of f
     datalength = readvariablelength(f)
